fix: deduplicate THEMATIC_BREAK_LINE regex and move slide utilities to parser
Export THEMATIC_BREAK_LINE from parser.ts instead of redefining it in
MarkdownEditor.tsx. Move getSlideIndexAtPosition and findSlideOffset to
parser.ts since they operate on the same slide-splitting logic, ensuring
a single source of truth for slide boundary detection.

Closes #16
